requirements.txt
arrow==0.13.0
boto==2.49.0
Django==1.11.18
django-debug-toolbar==1.11
django-decorator-include==1.4.1
python-memcached==1.59
django-cache-machine==1.1.0
django-js-reverse==0.8.2
django-filter==1.1
djangorestframework==3.9.0
djangorestframework-bulk==0.2.1
djangorestframework-camel-case==0.2.0
celery==3.1.19
redis==2.10.5
enum34==1.1.6
flake8==3.6.0
Jinja2==2.10
oEmbed==0.1.2
python-oembed==0.2.4
pytils==0.3
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
simplejson==3.16.0
drf-writable-nested==0.5.1
pendulum==2.0.4
simple-salesforce==0.74.2
selenium==3.13.0
django-anymail[mailgun]==2.0
openpyxl==2.5.4
xlwt==1.3.0
